+++ src/containerapp/azext_containerapp/custom.py
@@ -458,6 +458,7 @@ def create_containerapp(cmd,
                         env_vars=None,
                         cpu=None,
                         memory=None,
+                        gpu=None,
                         registry_server=None,
                         registry_user=None,
                         registry_pass=None,
@@ -542,6 +543,7 @@ def update_containerapp_logic(cmd,
                               target_label=None,
                               cpu=None,
                               memory=None,
+                              gpu=None,
                               revision_suffix=None,
                               startup_command=None,
                               args=None,
@@ -610,6 +612,7 @@ def update_containerapp(cmd,
                         target_label=None,
                         cpu=None,
                         memory=None,
+                        gpu=None,
                         revision_suffix=None,
                         startup_command=None,
                         args=None,
@@ -652,6 +655,7 @@ def update_containerapp(cmd,
                                      target_label=target_label,
                                      cpu=cpu,
                                      memory=memory,
+                                     gpu=gpu,
                                      revision_suffix=revision_suffix,
                                      startup_command=startup_command,
                                      args=args,
@@ -911,6 +915,7 @@ def create_containerappsjob(cmd,
                             env_vars=None,
                             cpu=None,
                             memory=None,
+                            gpu=None,
                             registry_server=None,
                             registry_user=None,
                             registry_pass=None,
@@ -967,6 +972,7 @@ def update_containerappsjob(cmd,
                             remove_all_env_vars=False,
                             cpu=None,
                             memory=None,
+                            gpu=None,
                             startup_command=None,
                             args=None,
                             scale_rule_metadata=None,
@@ -1449,7 +1455,7 @@ def create_containerapps_from_compose(cmd,  # pylint: disable=R0914
                                                                        resolve_replicas_from_service,
                                                                        resolve_environment_from_service,
                                                                        resolve_secret_from_service)
-    from ._compose_utils import validate_memory_and_cpu_setting
+    from ._compose_utils import resolve_gpu_configuration_from_service, validate_memory_and_cpu_setting
 
     # Validate managed environment
     parsed_managed_env = parse_resource_id(managed_env)
@@ -1494,6 +1500,7 @@ def create_containerapps_from_compose(cmd,  # pylint: disable=R0914
             resolve_memory_configuration_from_service(service),
             managed_environment
         )
+        gpu = resolve_gpu_configuration_from_service(service)
         replicas = resolve_replicas_from_service(service)
         environment = resolve_environment_from_service(service)
         secret_vars, secret_env_ref = resolve_secret_from_service(service, parsed_compose_file.secrets)
@@ -1540,6 +1547,7 @@ def create_containerapps_from_compose(cmd,  # pylint: disable=R0914
                                 args=startup_args,
                                 cpu=cpu,
                                 memory=memory,
+                                gpu=gpu,
                                 env_vars=environment,
                                 secrets=secret_vars,
                                 min_replicas=replicas,